Metadata features report fixes.
